Create Difference between the length of the rectangles
Alen and Tim both own a tennis grass court and they decide to mow the lawn in and around the court which will cost them Rs.5 per square feet. Given the amount they spent to mow the lawn and the width of the court, find the difference between the length of the courts.
input format:
First line will contain the amount spent (in Rs) by Alen and Tim separated by space. Second line will contain the width (in feet) of the courts of Alen and Tim separated by space.
Output format
The value (in feet) which is the difference between the length of the courts rounded off upto two decimal points.
Example Input/Output 1: Input: 100000 80000 100 80 Output: 0 Explanation: Area of Alen's court = 100000/5 = 20000 sq.ft. Length = 20000/100 = 200 Area of Tim's court = 80000/5 = 16000 sq.ft. Length = 16000/80 = 200 Hence the difference = 200-200 = 0 which when rounded off to decimal places is 0.00